feat: move TileManager and TileCreator to own files
